requirements.txt
requests
python-dotenv
streamlit
mysql-connector-python